+++ Auk/Auk.swift
@@ -30,7 +30,8 @@ public class Auk {
 
   */
   public var settings = AukSettings()
-
+  
+    
   /**
 
   Shows a local image in the scroll view.
@@ -62,6 +63,48 @@ public class Auk {
       AukPageVisibility.tellPagesAboutTheirVisibility(scrollView, settings: settings)
     }
   }
+  
+  /**
+   
+   Updates existing image in the scroll view.
+   
+   - parameter pageIndex: the index of the image to change.
+   - parameter image: Image to be shown in the scroll view.
+   - parameter accessibilityLabel: Text describing the image that will be spoken in accessibility mode.
+   For example: "Picture of a pony standing in a flower pot.".
+   
+   */
+  public func updateAt(pageIndex:Int, image: UIImage, accessibilityLabel: String? = nil) {
+    if let scrollView = scrollView,
+      page = AukScrollViewContent.pageAt(pageIndex, scrollView: scrollView) {
+      
+      page.prepareForReuse()
+      page.accessibilityLabel = accessibilityLabel
+      page.show(image: image, settings: settings)
+    }
+  }
+  
+  /**
+   
+   Downloads a remote image and updates the selected page in the scroll view. Use `Moa.settings.cache` property to configure image caching.
+   
+   - parameter pageIndex: the index of the image to change.
+   - parameter url: Url of the image to be shown.
+   - parameter accessibilityLabel: Text describing the image that will be spoken in accessibility mode.
+   For example: "Picture of a pony standing in a flower pot.".
+   
+   */
+  public func updateAt(pageIndex: Int, url: String, accessibilityLabel: String? = nil) {
+    if let scrollView = scrollView,
+      page = AukScrollViewContent.pageAt(pageIndex, scrollView: scrollView) {
+      
+      page.prepareForReuse()
+      page.accessibilityLabel = accessibilityLabel
+      page.show(url: url, settings: settings)
+      
+      AukPageVisibility.tellPagesAboutTheirVisibility(scrollView, settings: settings)
+    }
+  }
 
   /**
 
@@ -85,7 +128,7 @@ public class Auk {
   This function can be used for animating the scroll view content during orientation change. It is called in viewWillTransitionToSize and inside animateAlongsideTransition animation block.
 
       override func viewWillTransitionToSize(size: CGSize,
-      withTransitionCoordinator coordinator: UIViewControllerTransitionCoordinator) {
+        withTransitionCoordinator coordinator: UIViewControllerTransitionCoordinator) {
 
         super.viewWillTransitionToSize(size, withTransitionCoordinator: coordinator)
 
@@ -128,7 +171,7 @@ public class Auk {
 
   */
   public func scrollToNextPage(cycle cycle: Bool, animated: Bool) {
-    if let scrollView = scrollView {
+    if let scrollView = scrollView, currentPageIndex = currentPageIndex {
       AukScrollTo.scrollToNextPage(scrollView, cycle: cycle, animated: animated,
         currentPageIndex: currentPageIndex, numberOfPages: numberOfPages)
     }
@@ -152,7 +195,7 @@ public class Auk {
 
   */
   public func scrollToPreviousPage(cycle cycle: Bool, animated: Bool) {
-    if let scrollView = scrollView {
+    if let scrollView = scrollView, currentPageIndex = currentPageIndex {
       AukScrollTo.scrollToPreviousPage(scrollView, cycle: cycle, animated: animated,
         currentPageIndex: currentPageIndex, numberOfPages: numberOfPages)
     }
@@ -173,7 +216,10 @@ public class Auk {
     }
 
     pageIndicatorContainer?.updateNumberOfPages(numberOfPages)
-    pageIndicatorContainer?.updateCurrentPage(currentPageIndex)
+    
+    if let currentPageIndex = currentPageIndex {
+      pageIndicatorContainer?.updateCurrentPage(currentPageIndex)
+    }
   }
 
   /// Returns the current number of pages.
@@ -202,10 +248,12 @@ public class Auk {
 
   /**
 
-  Returns the current page index. If pages are being scrolled and there are two of them on screen the page index will indicate the page that occupies bigger portion of the screen at the moment.
+  Returns the current page index. If pages are being scrolled and there are two of them on screen the page index will indicate the page that occupies bigger portion of the screen at the moment. Returns nil if there are no pages. If scrolled way to the left or right beyond the pages it will return zero or the last index respectively.
 
   */
-  public var currentPageIndex: Int {
+  public var currentPageIndex: Int? {
+    if numberOfPages == 0 { return nil }
+    
     if let scrollView = scrollView {
       let width = Double(scrollView.bounds.size.width)
       let offset = Double(scrollView.contentOffset.x)
@@ -215,13 +263,15 @@ public class Auk {
       // Page # 0 is the rightmost in the right-to-left language layout
       if RightToLeft.isRightToLeft(scrollView) {
         value = numberOfPages - value - 1
-        if value < 0 { value = 0 }
       }
+      
+      if value < 0 { value = 0 }
+      if value > numberOfPages - 1 { value = numberOfPages - 1 }
       
       return value
     }
 
-    return 0
+    return nil
   }
 
   /**
@@ -295,7 +345,7 @@ public class Auk {
     scrollView?.showsHorizontalScrollIndicator = settings.showsHorizontalScrollIndicator
     scrollView?.pagingEnabled = settings.pagingEnabled
   }
-
+    
   /// Create a page, add it to the scroll view content and layout.
   private func createPage(accessibilityLabel: String? = nil) -> AukPage {
     let page = AukPage()
@@ -316,8 +366,10 @@ public class Auk {
     }
 
     pageIndicatorContainer?.updateNumberOfPages(numberOfPages)
-    pageIndicatorContainer?.updateCurrentPage(currentPageIndex)
     
+    if let currentPageIndex = currentPageIndex {
+      pageIndicatorContainer?.updateCurrentPage(currentPageIndex)
+    }
 
     return page
   }
@@ -325,7 +377,10 @@ public class Auk {
   func onScroll() {
     if let scrollView = scrollView {
       AukPageVisibility.tellPagesAboutTheirVisibility(scrollView, settings: settings)
-      pageIndicatorContainer?.updateCurrentPage(currentPageIndex)
+      
+      if let currentPageIndex = currentPageIndex {
+        pageIndicatorContainer?.updateCurrentPage(currentPageIndex)
+      }
     }
   }